Add pyproject.toml to modernise packaging to PEP621
I've copied over all the relevant metadata from the older setup.cfg
and setup.py files into the new pyproject.toml. The new packaging
also installs command line scripts and requirements which are
read from the requirments.txt.
I added a python_requires to be >=3.7 - this was mainly chosen
since this aligns with detectron2's requirements as will be needed
for the aerial-segmenataion install.