Improve tile merging and get rid of gdal dependency.
Tiling is now done by the framework in the `aerial-conversion` via a
call to `save_tiles` which now take `tile_overlap` as a percentage.
Merging is done via the new function `merge_mask` that sums the
weighted masks in tile overlap regions and only accepts pixels that
are >50% masked.

Also add an option to name the temporary tile and mask directories, which
is useful for annotating large datasets in parrallel.